Revert "Rerenders fix, better folder structure and naming"
