+++ src/app/page.tsx
@@ -5,7 +5,6 @@ import { OGameStatus, useGameStore } from "@/features/playzone/model";
 import { Button } from "@/shared/ui/button";
 import { AnimatePresence, LazyMotion, m } from "framer-motion";
 import { Loader2 } from "lucide-react";
-import { useShallow } from "zustand/react/shallow";
 const checkWordUrl = "https://api.dictionaryapi.dev/api/v2/entries/en/";
 
 const variants = {
@@ -14,12 +13,12 @@ const variants = {
 };
 export default function Home() {
   const { gameStatus, error, levelColor, fetchLevelWords } = useGameStore(
-    useShallow((state) => ({
+    (state) => ({
       gameStatus: state.gameStatus,
       error: state.error,
       levelColor: state.levelColor,
       fetchLevelWords: state.fetchLevelWords,
-    })),
+    }),
   );
 
   return (
