+++ src/features/playzone/index.ts
@@ -1,1 +1,1 @@
-export { Playzone } from './playzone';
+export { Playzone } from "./playzone";
