+++ src/features/playzone/model/index.ts
@@ -1,3 +1,2 @@
-export { useGameStore } from './game-store';
-export { useKeyboardControlStore } from './keyboard-control-store';
-export { OGameStatus, type TGameStatus, type WordData, type WordDataFetched } from './types';
+export { useKeyboardControlStore } from "./keyBoardControlStore";
+export { useGameStore, OGameStatus } from "./gameStore";
