--- src/features/playzone/model/types.ts
@@ -1,20 +1,0 @@
-export interface WordDataFetched {
-  word: string;
-  score: number;
-  defs?: string[];
-}
-
-export interface WordData {
-  [word: string]: {
-    score: number;
-    defs?: string[];
-  };
-}
-
-export const OGameStatus = {
-  NOT_STARTED: 'NOT_STARTED',
-  LOADING: 'LOADING',
-  ON_LVL: 'ON_LVL',
-  AFTER_LVL: 'AFTER_LVL',
-} as const;
-export type TGameStatus = (typeof OGameStatus)[keyof typeof OGameStatus];
