--- src/features/playzone/ui/index.ts
@@ -1,3 +1,0 @@
-export { default as Cell } from './cell';
-export { default as CellButton } from './cell-button';
-export { containerVariants } from './animation-variants';
